requirements.txt
mypy==1.2.0
isort==5.12.0
ipykernel==6.22.0
black==23.3.0
pytorch-lightning==2.0.2
scikit-learn==1.2.2
einops==0.6.1
opencv-python==4.7.0.72
opencv-transforms==0.0.6
dpath==2.1.5
unflatten==0.1.1
lpips==0.1.4
Levenshtein==0.21.0
cachetools==5.3.0
pytesseract==0.3.10
pytorch_msssim==0.2.1
timm==0.6.13
tensorboard==2.12.3
ipywidgets==8.0.6
torchinfo==1.8.0
icecream==2.1.3
types-cachetools==5.3.0.5
gdown==4.7.1
h5py==3.10.0
matplotlib==3.8.2
scikit-image==0.22.0
pandas==1.1.5
plotly==5.6.0
pre-commit==3.6.0